move files from middleware add filed to User move files from middleware to middlewares, add filed password to User
