+++ src/db.ts
@@ -12,6 +12,7 @@ interface User {
     phone?: string;
     avatar?: string;
     emailConfirmed?: boolean;
+    password: string;
 }
 
 // 2. Create a Schema corresponding to the document interface.
@@ -37,7 +38,8 @@ const userSchema = new Schema<User>({
         lowercase: true
     },
     avatar: String,
-    emailConfirmed: Boolean
+    emailConfirmed: Boolean,
+    password: String,
 });
 
 // 3. Create a Model.
